+++ lib/kitchen/terraform/breaking.rb
@@ -14,29 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "rubygems"
 require "kitchen/terraform"
 
-# The version of the kitchen-terraform gem.
-class ::Kitchen::Terraform::Version
-  def assign_specification_version(specification:)
-    specification.version = @version.to_s
-    self
-  end
-
-  def assign_plugin_version(configurable_class:)
-    configurable_class.plugin_version @version.to_s
-    self
-  end
-
-  def if_satisfies(requirement:)
-    yield if requirement.satisfied_by? @version
-    self
-  end
-
-  private
-
-  def initialize(version: "3.3.0")
-    @version = ::Gem::Version.new version
-  end
+# The namespace for objects with breaking changes.
+module ::Kitchen::Terraform::Breaking
 end