+++ lib/kitchen/terraform/configurable.rb
@@ -16,6 +16,7 @@
 
 require "kitchen"
 require "kitchen/terraform"
+require "kitchen/terraform/kitchen_instance"
 require "kitchen/terraform/version"
 
 # Refinements to Kitchen::Configurable.
@@ -24,33 +25,36 @@ require "kitchen/terraform/version"
 module ::Kitchen::Terraform::Configurable
   # A callback to define the plugin version which is invoked when this module is included in a plugin class.
   #
-  # @return [void]
+  # @return [self]
   def self.included(configurable_class)
-    configurable_class.plugin_version ::Kitchen::Terraform::VERSION
+    ::Kitchen::Terraform::Version
+      .new
+      .assign_plugin_version configurable_class: configurable_class
+
+    self
   end
 
   # Alternative implementation of Kitchen::Configurable#finalize_config! which validates the configuration before
   # attempting to expand paths.
   #
   # @note this method should be removed when Test Kitchen: Issue #1229 is solved.
-  # @param instance [::Kitchen::Instance] an associated instance.
+  # @param kitchen_instance [::Kitchen::Instance] an associated Test Kitchen instance.
   # @return [self] itself, for use in chaining.
   # @raise [::Kitchen::ClientError] if the instance is nil.
   # @see https://github.com/test-kitchen/test-kitchen/blob/v1.16.0/lib/kitchen/configurable.rb#L46
   #   Kitchen::Configurable#finalize_config!
   # @see https://github.com/test-kitchen/test-kitchen/issues/1229 Test Kitchen: Issue #1229
-  def finalize_config!(instance)
-    instance or
+  def finalize_config!(kitchen_instance)
+    kitchen_instance or
       raise(
         ::Kitchen::ClientError,
         "Instance must be provided to #{self}"
       )
 
-    @instance = instance
+    @instance = ::Kitchen::Terraform::KitchenInstance.new kitchen_instance: kitchen_instance
     validate_config!
     expand_paths!
     load_needed_dependencies!
-
     self
   end
 end