+++ CHANGELOG.md
@@ -8,6 +8,12 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+* Deprecating support for concurrency with the following commands:
+  `create`, `converge`, `setup`, and `destroy`; these commands invoke
+  Terraform in a manner which is not safe for concurrency
+
 ## [3.3.0] - 2018-03-22
 
 ### Added